Match: Move DB logic into repository.ts
