+++ services/match/src/controllers/matchRequestController.ts
@@ -1,8 +1,12 @@
 import { Request, Response } from 'express';
 import { handleBadRequest, handleInternalError, handleNotFound, handleSuccess } from '../utils/helpers';
-import { MatchRequest } from '../models/matchRequestModel';
 import { isValidObjectId } from 'mongoose';
 import { createMatchRequestSchema, updateMatchRequestSchema } from '../validation/matchRequestValidation';
+import {
+    createMatchRequest as _createMatchRequest,
+    findMatchRequestAndUpdate,
+    findMatchRequestAndDelete,
+} from '../models/repository';
 
 /**
  * Creates a match request.
@@ -10,15 +14,15 @@ import { createMatchRequestSchema, updateMatchRequestSchema } from '../validatio
  * @param res
  */
 export const createMatchRequest = async (req: Request, res: Response) => {
-    const user = req.user;
     const { error, value } = createMatchRequestSchema.validate(req.query);
     if (error) {
         return handleBadRequest(res, error.message);
     }
 
+    const userId = req.user.id;
+    const { topics, difficulty } = value;
     try {
-        const matchRequest = new MatchRequest({ userId: user.id, ...value });
-        await matchRequest.save();
+        const matchRequest = await _createMatchRequest(userId, topics, difficulty);
         handleSuccess(res, 201, 'Match request created successfully', matchRequest);
     } catch (error) {
         console.error('Error in createMatchRequest:', error);
@@ -32,15 +36,16 @@ export const createMatchRequest = async (req: Request, res: Response) => {
  * @param res
  */
 export const updateMatchRequest = async (req: Request, res: Response) => {
-    const id = req.params.id;
-    const userId = req.user.id;
     const { error, value } = updateMatchRequestSchema.validate(req.query);
     if (error) {
         return handleBadRequest(res, error.message);
     }
 
+    const id = req.params.id;
+    const userId = req.user.id;
+    const { topics, difficulty } = value;
     try {
-        const matchRequest = await MatchRequest.findOneAndUpdate({ _id: id, userId }, value);
+        const matchRequest = await findMatchRequestAndUpdate(id, userId, topics, difficulty);
         handleSuccess(res, 201, 'Match request update successfully', matchRequest);
     } catch (error) {
         console.error('Error in updateMatchRequest:', error);
@@ -62,7 +67,7 @@ export const deleteMatchRequest = async (req: Request, res: Response) => {
     }
 
     try {
-        const matchRequest = await MatchRequest.findOneAndDelete({ _id: id, userId });
+        const matchRequest = await findMatchRequestAndDelete(id, userId);
         if (!matchRequest) {
             return handleNotFound(res, `Request ${id} not found`);
         }
